separate assets module
